base finish

+++ app/Http/Controllers/ServiceController.php
@@ -10,6 +10,7 @@ use App\Models\Service;
 use Illuminate\Database\RecordsNotFoundException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Validator;
 
 class ServiceController extends Controller
 {
@@ -54,11 +55,20 @@ class ServiceController extends Controller
 
     public function bulkUpdate(Request $request): JsonResponse
     {
-        //TODO request controlü yapılacak.
         try {
         $data = $request->all();
         foreach ($data as $item)
         {
+            $validator = Validator::make($item, [
+                'name' => 'required',
+                'price' => 'required|integer',
+                'cost' => 'required|integer|lt:price'
+            ]);
+
+            if($validator->fails()){
+               throw new \Exception($validator->errors()->first(), 404);
+            }
+
             $this->serviceRepository->delete($item['service_id']);
             $update_data = ['name' => $item['name'], 'price' => $item['price'], 'cost' => $item['cost'] ];
